refactor: extract pagination constants and resolvers from resumes.ts
Move 12 constants and 8 resolver functions into lib/resumes_pagination.ts,
reducing resumes.ts by ~76 net lines. All symbols re-exported for backward
compatibility.
